+++ test-app/android-views/src/main/kotlin/com/example/redwood/testapp/android/views/TestAppActivity.kt
@@ -146,7 +146,11 @@ class TestAppActivity : ComponentActivity() {
         manifestUrl = manifestUrlFlow,
         hostApi = RealHostApi(httpClient),
       ),
-      eventListenerFactory = { _, _ -> appEventListener },
+      eventListenerFactory = object : EventListener.Factory {
+        override fun create(app: TreehouseApp<*>, manifestUrl: String?) = appEventListener
+        override fun close() {
+        }
+      },
     )
 
     treehouseApp.start()